refactor(authSample): changes auth flow and adds tests
