+++ samples/auth-user-sample/.eslintrc.yml
@@ -1,0 +1,4 @@
+---
+rules:
+  no-process-exit: off
+  node/no-unpublished-require: off
