+++ samples/auth-user-sample/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "nodejs-docs-samples-bigquery",
+  "name": "auth-user-sample",
   "files": [
     "*.js",
     "resources/"
@@ -16,11 +16,14 @@
   },
   "dependencies": {
     "@google-cloud/bigquery": "^4.4.0",
-    "@google-cloud/storage": "^4.0.0"
+    "google-auth-library": "^5.5.1",
+    "readline": "^1.3.0",
+    "sinon-chai": "^3.3.0"
   },
   "devDependencies": {
     "chai": "^4.2.0",
     "mocha": "^6.0.0",
+    "proxyquire": "^2.1.3",
     "sinon": "^7.5.0",
     "uuid": "^3.3.0"
   }
