--- samples/config.keys.json
@@ -1,4 +1,0 @@
-{
-    "username": "your_email",
-    "password": "your_password"
-}
