--- samples/oauth2.keys.json
@@ -1,1 +1,0 @@
-{"installed":{"client_id":"client_id","project_id":"project_id","auth_uri":"auth_uri","token_uri":"token_uri","auth_provider_x509_cert_url":"auth_provider","client_secret":"client_secret","redirect_uris":["redirect_uri"]}}
